app/rides/views/review.py
from rest_framework import viewsets

from rides.models import Review
from rides.serializers import ReviewSerializer


class ReviewViewSet(viewsets.ModelViewSet):
    queryset = Review.objects.all()
    serializer_class = ReviewSerializer