app/requirements.txt
django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
djoser==2.3.1
django-filter==23.5
django-environ==0.11.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
psycopg2-binary~=2.0
gunicorn==22.0.0
